勾股OA5.5版本更新

+++ app/home/controller/Role.php
@@ -1,9 +1,15 @@
 <?php
 /**
- * @copyright Copyright (c) 2021 勾股工作室
- * @license https://opensource.org/licenses/GPL-2.0
- * @link https://www.gougucms.com
- */
++-----------------------------------------------------------------------------------------------
+* GouGuOPEN [ 左手研发，右手开源，未来可期！]
++-----------------------------------------------------------------------------------------------
+* @Copyright (c) 2021~2024 http://www.gouguoa.com All rights reserved.
++-----------------------------------------------------------------------------------------------
+* @Licensed 勾股OA，开源且可免费使用，但并不是自由软件，未经授权许可不能去除勾股OA的相关版权信息
++-----------------------------------------------------------------------------------------------
+* @Author 勾股工作室 <hdm58@qq.com>
++-----------------------------------------------------------------------------------------------
+*/
 
 declare (strict_types = 1);
 
@@ -26,11 +32,8 @@ class Role extends BaseController
             if (!empty($param['keywords'])) {
                 $where[] = ['id|title|desc', 'like', '%' . $param['keywords'] . '%'];
             }
-            $rows = empty($param['limit']) ? get_config('app . page_size') : $param['limit'];
-            $group = AdminGroup::where($where)
-                ->order('create_time asc')
-                ->paginate($rows, false, ['query' => $param]);
-            return table_assign(0, '', $group);
+			$list = Db::name('AdminGroup')->where($where)->order('create_time asc')->select();
+            return to_assign(0, '', $list);
         } else {
             return view();
         }
@@ -42,7 +45,19 @@ class Role extends BaseController
         $param = get_params();
         if (request()->isAjax()) {
             $ruleData = isset($param['rule']) ? $param['rule'] : 0;
+            $layoutData = isset($param['layout']) ? $param['layout'] : 0;
+            $menuData = isset($param['mobile_menu']) ? $param['mobile_menu'] : 0;
+            $barData = isset($param['mobile_bar']) ? $param['mobile_bar'] : 0;
+			if($ruleData==0){
+				return to_assign(1, '权限节点至少选择一个');
+			}
+			if($layoutData==0){
+				return to_assign(1, '首页展示模块至少选择一个');
+			}
             $param['rules'] = implode(',', $ruleData);
+            $param['layouts'] = implode(',', $layoutData);
+            $param['mobile_menu'] = implode(',', $menuData);
+            $param['mobile_bar'] = implode(',', $barData);
             if (!empty($param['id']) && $param['id'] > 0) {
                 try {
                     validate(GroupCheck::class)->scene('edit')->check($param);
@@ -68,19 +83,75 @@ class Role extends BaseController
             }
             //清除菜单\权限缓存
             clear_cache('adminMenu');
+            clear_cache('MobileRules');
             return to_assign();
         } else {
             $id = isset($param['id']) ? $param['id'] : 0;
             $rule = admin_rule();
+			$layouts = get_config('layout');
+			$mobile_bar = Db::name('MobileBar')->where([['status','=',1]])->field('id,url,title,icon')->select()->toArray();
+			$mobile_menu = Db::name('MobileTypes')->where(['status'=>1])->select()->toArray();		
+
+			
             if ($id > 0) {
                 $rules = admin_group_info($id);
                 $role_rule = create_tree_list(0, $rule, $rules);
-                $role = Db::name('AdminGroup')->where(['id' => $id])->find();
+                $role = Db::name('AdminGroup')->where(['id' => $id])->find();				
                 View::assign('role', $role);
+				
+				$layout_selected = explode(',', $role['layouts']);
+				foreach ($layouts as $key =>&$vo) {
+					if (!empty($layout_selected) and in_array($vo['id'], $layout_selected)) {
+						$vo['checked'] = true;
+					} else {
+						$vo['checked'] = false;
+					}
+				}
+				
+				$mobile_bar_selected = explode(',', $role['mobile_bar']);
+				foreach ($mobile_bar as $key =>&$vo) {
+					if (!empty($mobile_bar_selected) and in_array($vo['id'], $mobile_bar_selected)) {
+						$vo['checked'] = true;
+					} else {
+						$vo['checked'] = false;
+					}
+				}
+				
+				$mobile_menu_selected = explode(',', $role['mobile_menu']);
+				foreach ($mobile_menu as &$row) {
+					$list = Db::name('MobileMenu')->where([['types','=',$row['id']],['status','=',1]])->select()->toArray();
+					foreach ($list as $key =>&$vo) {
+						if (!empty($mobile_menu_selected) and in_array($vo['id'], $mobile_menu_selected)) {
+							$vo['checked'] = true;
+						} else {
+							$vo['checked'] = false;
+						}
+					}	
+					$row['list'] = $list;
+				}
+
             } else {
                 $role_rule = create_tree_list(0, $rule, []);
-            }
-            View::assign('role_rule', $role_rule);
+				foreach ($layouts as $key =>&$vo) {
+					$vo['checked'] = false;
+				}
+				
+				foreach ($mobile_bar as $key =>&$vo) {
+					$vo['checked'] = false;
+				}
+				
+				foreach ($mobile_menu as &$row) {
+					$list = Db::name('MobileMenu')->where([['types','=',$row['id']],['status','=',1]])->select()->toArray();
+					foreach ($list as $key =>&$vo) {
+						$vo['checked'] = false;
+					}	
+					$row['list'] = $list;
+				}
+            }	
+            View::assign('role_rule', $role_rule);			
+            View::assign('layout', $layouts);
+            View::assign('mobile_bar', $mobile_bar);
+            View::assign('mobile_menu', $mobile_menu);
             View::assign('id', $id);
             return view();
         }
@@ -89,15 +160,23 @@ class Role extends BaseController
     //删除
     public function delete()
     {
-        $id = get_params("id");
-        if ($id == 1) {
-            return to_assign(1, "该组是系统所有者，无法删除");
-        }
-        if (Db::name('AdminGroup')->delete($id) !== false) {
-            add_log('delete', $id, []);
-            return to_assign(0, "删除角色成功");
+        if (request()->isDelete()) {
+            $id = get_params("id");
+            if ($id == 1) {
+                return to_assign(1, "该组是系统所有者，无法删除");
+            }
+            $count = Db::name('PositionGroup')->where(["group_id" => $id])->count();
+            if ($count > 0) {
+                return to_assign(1, "该权限组还在使用，请去除使用者关联再删除");
+            }
+            if (Db::name('AdminGroup')->delete($id) !== false) {
+                add_log('delete', $id, []);
+                return to_assign(0, "删除权限组成功");
+            } else {
+                return to_assign(1, "删除失败");
+            }
         } else {
-            return to_assign(1, "删除失败");
+            return to_assign(1, "错误的请求");
         }
     }
 }
