+++ app/home/controller/Rule.php
@@ -1,9 +1,15 @@
 <?php
 /**
- * @copyright Copyright (c) 2021 勾股工作室
- * @license https://opensource.org/licenses/GPL-2.0
- * @link https://www.gougucms.com
- */
++-----------------------------------------------------------------------------------------------
+* GouGuOPEN [ 左手研发，右手开源，未来可期！]
++-----------------------------------------------------------------------------------------------
+* @Copyright (c) 2021~2024 http://www.gouguoa.com All rights reserved.
++-----------------------------------------------------------------------------------------------
+* @Licensed 勾股OA，开源且可免费使用，但并不是自由软件，未经授权许可不能去除勾股OA的相关版权信息
++-----------------------------------------------------------------------------------------------
+* @Author 勾股工作室 <hdm58@qq.com>
++-----------------------------------------------------------------------------------------------
+*/
 
 declare (strict_types = 1);
 
@@ -21,12 +27,13 @@ class Rule extends BaseController
     {
         if (request()->isAjax()) {
             $rule = Db::name('adminRule')
-			->field('a.*,m.title as module_title')
-			->alias('a')
-			->leftJoin('adminModule m','a.module = m.name')
-			->order('a.sort asc,a.id asc')
-			->select();
-            return to_assign(0, '', $rule);
+                ->field('a.*,m.title as module_title')
+                ->alias('a')
+                ->leftJoin('adminModule m', 'a.module = m.name')
+                ->order('a.sort asc,a.id asc')
+                ->select();
+			$list = generateTree($rule);
+            return to_assign(0, '', $list);
         } else {
             return view();
         }
@@ -37,7 +44,7 @@ class Rule extends BaseController
     {
         $param = get_params();
         if (request()->isAjax()) {
-			$param['src'] = preg_replace('# #','',$param['src']);
+            $param['src'] = preg_replace('# #', '', $param['src']);
             if ($param['id'] > 0) {
                 try {
                     validate(RuleCheck::class)->scene('edit')->check($param);
@@ -56,7 +63,7 @@ class Rule extends BaseController
                 }
                 $param['create_time'] = time();
                 $rid = Db::name('AdminRule')->strict(false)->field(true)->insertGetId($param);
-                //自动为系统所有者管理组分配新增的节点
+                //自动为系统超级管理员角色组分配新增的节点
                 $group = Db::name('AdminGroup')->find(1);
                 if (!empty($group)) {
                     $newGroup['id'] = 1;
@@ -71,8 +78,8 @@ class Rule extends BaseController
         } else {
             $id = isset($param['id']) ? $param['id'] : 0;
             $pid = isset($param['pid']) ? $param['pid'] : 0;
-            if($id>0){
-                $detail = Db::name('AdminRule')->where('id',$id)->find();
+            if ($id > 0) {
+                $detail = Db::name('AdminRule')->where('id', $id)->find();
                 View::assign('detail', $detail);
             }
             View::assign('id', $id);
@@ -83,17 +90,21 @@ class Rule extends BaseController
     //删除
     public function delete()
     {
-        $id = get_params("id");
-        $count = Db::name('AdminRule')->where(["pid" => $id])->count();
-        if ($count > 0) {
-            return to_assign(1, "该节点下还有子节点，无法删除");
-        }
-        if (Db::name('AdminRule')->delete($id) !== false) {
-            clear_cache('adminRules');
-            add_log('delete', $id, []);
-            return to_assign(0, "删除节点成功");
+        if (request()->isDelete()) {
+            $id = get_params("id");
+            $count = Db::name('AdminRule')->where(["pid" => $id])->count();
+            if ($count > 0) {
+                return to_assign(1, "该节点下还有子节点，无法删除");
+            }
+            if (Db::name('AdminRule')->delete($id) !== false) {
+                clear_cache('adminRules');
+                add_log('delete', $id, []);
+                return to_assign(0, "删除节点成功");
+            } else {
+                return to_assign(1, "删除失败");
+            }
         } else {
-            return to_assign(1, "删除失败");
+            return to_assign(1, "错误的请求");
         }
     }
 }
